Move index into its own views file
