+++ core/views/__init__.py
@@ -13,3 +13,5 @@ from .stateio import *
 from .core_background_tasks import *
 from .export import *
 from .external_background_tasks import *
+from .analysis import *
+from .index import *
